script 3

+++ more_malloc_free/3-array_range.c
@@ -7,19 +7,19 @@
  */
 int *array_range(int min, int max)
 {
-    int *p;
-    int i;
+	int *p;
+	int i;
 
-    if (max < min)
-        return (NULL);
+	if (max < min)
+		return (NULL);
 
-    p = (int)malloc(sizeof(int) * (max - min + 1));
+	p = malloc(sizeof(int) * (max - min + 1));
 
-    if (p == NULL)
-        return (NULL);
+	if (p == NULL)
+		return (NULL);
 
-    for (i = 0; i <= max - min + 1; i++)
-        p[i] = min + i;
+	for (i = 0; i <= max - min + 1; i++)
+		p[i] = min + i;
 
-    return (p);
+	return (p);
 }
